requirements.txt
anthropic>=0.40.0
feedparser>=6.0.0
pyyaml>=6.0
requests>=2.31.0
beautifulsoup4>=4.12.0
scikit-learn>=1.4.0
python-dotenv>=1.0.0
jinja2>=3.1.0
markupsafe>=2.1.0
